Codigos/2025/programacion IV/Tareas/5 migracion a KeyDB como almacenamiento en memoria/requirements.txt
redis
python-dotenv
